Feat: 알고리즘 계산 백분율 변경
백분율 = (해당 태그 문제 수 / 전체 푼 문제 수) × 100

+++ Koco/src/main/java/icet/koco/user/dto/UserAlgorithmStatsResponseDto.java
@@ -14,6 +14,6 @@ public class UserAlgorithmStatsResponseDto {
     public static class CategoryStat {
         private Long categoryId;
         private String categoryName;
-        private Double correctRate;
+        private Integer correctRate;
     }
 }
